% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCounties.R
\name{CalculateCounties}
\alias{CalculateCounties}
\title{#Get the total population in the selected region}
\usage{
CalculateCounties(IncludedCounties)
}
\description{
#Get the total population in the selected region
}
