% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCovid.R
\name{CalculateCovid}
\alias{CalculateCovid}
\title{Get total confirmed cases in the selected region}
\usage{
CalculateCovid(IncludedCounties, CovidConfirmedCases)
}
\description{
Get total confirmed cases in the selected region
}
