% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAD3-package.R
\docType{package}
\name{CHAD3-package}
\alias{CHAD3}
\alias{CHAD3-package}
\title{CHAD3: USAF COVID-19 Shiny App}
\description{
Data sets and support code for USAF COVID-19 Shiny app.
}
\author{
\strong{Maintainer}: Jason Freels \email{auburngrads@live.com}

Authors:
\itemize{
  \item Trey Pujats
  \item Nick Forrest
  \item Garret Alarcon
  \item Jeff Weir
}

}
\keyword{internal}
