% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_app.R
\name{covid19_app}
\alias{covid19_app}
\title{Runs the COVID19 App}
\usage{
covid19_app(
  app_name = "covid19",
  width = "100\%",
  height = "800px",
  launch.browser = TRUE,
  ...
)
}
\description{
Runs the COVID19 App
}
