% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lanl_data.R
\name{get_lanl_data}
\alias{get_lanl_data}
\title{Get data from Los Alamos National Lab}
\usage{
get_lanl_data(
  front = "https://covid-19.bsvgateway.org/forecast/us/files",
  middle = "/confirmed/",
  end = "_confirmed_quantiles_us.csv",
  date = Sys.Date() - 4
)
}
\arguments{
\item{front}{front of url}

\item{middle}{middle of url}

\item{end}{end of url}

\item{date}{Have to adjust the integer value based on the date of the most recent forecast file}
}
\description{
Get data from Los Alamos National Lab
}
