% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLocalDataTable.R, R/GetLocalDataTable2.R
\name{GetLocalDataTable}
\alias{GetLocalDataTable}
\title{Create Data Table for local statistics}
\usage{
GetLocalDataTable(IncludedCounties, CovidConfirmedCases, CovidDeaths)

GetLocalDataTable(IncludedCounties, CovidConfirmedCases, CovidDeaths)
}
\description{
Create Data Table for local statistics

Create Data Table for local statistics
}
