% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotOverlay.R
\name{PlotOverlay}
\alias{PlotOverlay}
\title{Create charts for projecting local health data}
\usage{
PlotOverlay(
  ChosenBase,
  IncludedCounties,
  IncludedHospitals,
  SocialDistance,
  DaysProjected,
  StatisticType
)
}
\description{
Create charts for projecting local health data
}
