% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotOverlay2.R
\name{PlotOverlay2}
\alias{PlotOverlay2}
\title{#' Create charts for projecting local health data}
\usage{
PlotOverlay2(
  ChosenBase,
  IncludedCounties,
  IncludedHospitals,
  SocialDistance,
  DaysProjected,
  StatisticType,
  AFBaseLocations,
  IHME_Model,
  CountyInfo,
  HospitalInfo,
  CovidConfirmedCases,
  CovidDeaths
)
}
\description{
#' Create charts for projecting local health data
}
\details{
EXTRA EXTRA Need to find a better way later to replace this 
         probably have the overlay function return a list with two objects. 
         need the data.frame from overlay in the report
}
