% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIAR_Model_Run.R
\name{SEIAR_Model_Run}
\alias{SEIAR_Model_Run}
\title{Function to establish the Army SEIAR model}
\usage{
SEIAR_Model_Run(
  num_init_cases,
  Pop.At.Risk,
  incub_period,
  latent_period,
  doubling,
  recovery_days,
  social_rate,
  hospital_rate,
  icu_rate,
  ventilated_rate,
  hospital_dur,
  icu_dur,
  ventilated_dur,
  n_days,
  secondary_cases = 2.5,
  distribution_e_to_a = 0.5
)
}
\arguments{
\item{num_init_cases}{}

\item{Pop.At.Risk}{}

\item{incub_period}{}

\item{latent_period}{}

\item{doubling}{}

\item{recovery_days}{}

\item{social_rate}{}

\item{hospital_rate}{}

\item{icu_rate}{}

\item{ventilated_rate}{}

\item{hospital_dur}{}

\item{icu_dur}{}

\item{ventilated_dur}{}

\item{n_days}{}

\item{secondary_cases}{}

\item{distribution_e_to_a}{}
}
\description{
Uses US Army models to create projections 
             for the local area around a base
}
\details{
This function allows us to pass though a simple 
         function to gather all statistics when we plot
}
