% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovidCasesCumChart.R
\name{CovidCasesCumChart}
\alias{CovidCasesCumChart}
\title{Cumulative COVID cases chart}
\usage{
CovidCasesCumChart(IncludedCounties)
}
\description{
Begin function to create chart of new cases for COVID-19 is a specified region around a specified base
}
