% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLocalChoro.R
\name{PlotLocalChoro}
\alias{PlotLocalChoro}
\title{Create plot of Covid Cases by County}
\usage{
PlotLocalChoro(IncludedCounties, ChosenBase, TypofPlot)
}
\description{
Create plot of Covid Cases by County
}
